Add an option to keep save files when you die.

+++ app/src/main/java/com/watabou/pixeldungeon/actors/hero/Hero.java
@@ -29,6 +29,7 @@ import com.watabou.pixeldungeon.Badges;
 import com.watabou.pixeldungeon.Bones;
 import com.watabou.pixeldungeon.Dungeon;
 import com.watabou.pixeldungeon.GamesInProgress;
+import com.watabou.pixeldungeon.PixelDungeon;
 import com.watabou.pixeldungeon.ResultDescriptions;
 import com.watabou.pixeldungeon.actors.Actor;
 import com.watabou.pixeldungeon.actors.Char;
@@ -1219,7 +1220,9 @@ public class Hero extends Char {
 			((Hero.Doom)cause).onDeath();
 		}
 		
-		Dungeon.deleteGame( Dungeon.hero.heroClass, true );
+		if (!PixelDungeon.keepSaveFiles()) {
+			Dungeon.deleteGame( Dungeon.hero.heroClass, true );
+		}
 	}
 	
 	@Override
